Bump version and increase number of rounds
Clutch cryptanalysis shows that the number of affected bytes in large messages
encrypted with Wring does not increase by a factor between 3 and 6 each round,
as I had expected, but increases by a factor that starts at 3.85 and drops
to just above 2 by six rounds. The number of rounds now increases by 1 for every
doubling of message size, rather than for every tripling as before.